+++ examples/apps/web-todomvc-ts/eslint.config.ts
@@ -1,0 +1,17 @@
+import browserConfig from "@inglorious/eslint-config/browser"
+import typescriptConfig from "@inglorious/eslint-config/typescript"
+import { defineConfig } from "eslint/config"
+
+export default defineConfig([
+  ...browserConfig,
+  ...typescriptConfig,
+  {
+    files: ["**/*.ts", "**/*.tsx"],
+    languageOptions: {
+      parserOptions: {
+        project: ["./tsconfig.app.json", "./tsconfig.node.json"],
+        tsconfigRootDir: import.meta.dirname,
+      },
+    },
+  },
+])