+++ examples/apps/web-todomvc-ts/src/app.ts
@@ -1,0 +1,9 @@
+import { type Api, html } from "@inglorious/web"
+
+export const app = {
+  render(api: Api) {
+    return html`${api.render("form")}${api.render("list")}${api.render(
+      "footer",
+    )}`
+  },
+}