+++ packages/web/types/lit-html.d.ts
@@ -1,0 +1,9 @@
+export { html, TemplateResult } from "lit-html"
+
+export { choose } from "lit-html/directives/choose"
+export { classMap } from "lit-html/directives/class-map"
+export { ref } from "lit-html/directives/ref"
+export { repeat } from "lit-html/directives/repeat"
+export { styleMap } from "lit-html/directives/style-map"
+export { unsafeHTML } from "lit-html/directives/unsafe-html"
+export { when } from "lit-html/directives/when"