+++ packages/web/types/store.d.ts
@@ -1,0 +1,14 @@
+export {
+  BaseEntity,
+  EntitiesState,
+  Event,
+  Middleware,
+  createStore,
+  TypesConfig,
+} from "@inglorious/store"
+
+export { createSelector } from "@inglorious/store/select"
+export { trigger } from "@inglorious/store/test"
+
+export { createDevtools } from "@inglorious/store/client/devtools"
+export { multiplayerMiddleware } from "@inglorious/store/client/multiplayer-middleware"