+++ examples/apps/web-router/CHANGELOG.md
@@ -1,5 +1,12 @@
 # lit-todomvc
 
+## 0.4.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@4.0.4
+
 ## 0.4.9
 
 ### Patch Changes