+++ examples/apps/web-select/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web-select
 
+## 0.2.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@4.0.4
+
 ## 0.2.7
 
 ### Patch Changes