+++ packages/create-app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inglorious/create-game
 
+## 0.3.3
+
+### Patch Changes
+
+- Add Inglorious Web TS example
+
 ## 0.3.2
 
 ### Patch Changes